+++ app/src/test/java/io/gnosis/safe/ui/transactions/details/viewdata/TransactionDetailsViewDataTest.kt
@@ -1,4 +1,4 @@
-package io.gnosis.safe.ui.transactions.details.models
+package io.gnosis.safe.ui.transactions.details.viewdata
 
 import io.gnosis.data.models.AddressInfo
 import io.gnosis.data.models.Safe
